# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "voice-ai api",
    "version": "0.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health",
        "description": "Liveness probe. Returns 200 with a static payload.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "Return the authenticated user's id and email.",
        "operationId": "me_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/livekit/token": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Livekit Token",
        "description": "Mint a LiveKit access token for the authenticated user.\n\nThe route is a thin adapter over :func:`core.livekit.issue_token`;\nthe token's `identity` is the Supabase user id and the room\ndefaults to ``user-{userId}`` when the client does not specify one.",
        "operationId": "livekit_token_livekit_token_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/LivekitTokenRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivekitTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/preferences": {
      "get": {
        "tags": [
          "preferences"
        ],
        "summary": "List Preferences",
        "description": "Return every stored preference for the authenticated user.\n\nThin adapter over :func:`core.preferences.list`. The user's bearer\ntoken is forwarded to the Supabase client so RLS policies (defined\nin ``0001_user_preferences.sql``) apply to the database query.",
        "operationId": "list_preferences_preferences_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/preferences/{key}": {
      "put": {
        "tags": [
          "preferences"
        ],
        "summary": "Upsert Preference",
        "description": "Upsert a single preference for the authenticated user.\n\nValidated against :func:`core.preferences.validate_preference`,\nwhich today only accepts the settings-page keys\n(:data:`core.preferences.SETTINGS_KEYS`). The free-form\n``set_preference`` agent tool remains available for keys outside\nthat catalogue \u2014 the settings page is intentionally narrower than\nthe agent's surface.\n\nReturns 400 with a structured error body on validation failure,\n401 without authentication, 200 on success.",
        "operationId": "upsert_preference_preferences__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Preference key, e.g. 'preferred_name' or 'voice'.",
              "title": "Key"
            },
            "description": "Preference key, e.g. 'preferred_name' or 'voice'."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceUpsertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations": {
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "List Conversations",
        "description": "Return the authenticated user's conversations, paginated.\n\nThin adapter over :func:`core.conversations.list_for_user`. RLS\nenforces user scoping at the database; the route just shapes the\nresponse. ``limit`` is bounded server-side so a hostile client\ncannot ask for the whole table.",
        "operationId": "list_conversations_conversations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "Get Conversation",
        "description": "Return one conversation with its full message log.\n\nReturns 404 when the conversation does not exist *or* belongs to\nanother user. Both surface here as ``None`` from the core layer\nbecause RLS makes \"not yours\" indistinguishable from \"not there\"\n\u2014 leaking the difference would itself be a privacy bug.",
        "operationId": "get_conversation_conversations__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Conversation UUID.",
              "title": "Conversation Id"
            },
            "description": "Conversation UUID."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConversationDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageItem"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "started_at"
        ],
        "title": "ConversationDetailResponse",
        "description": "Response payload for ``GET /conversations/{id}``."
      },
      "ConversationSummaryItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Conversation UUID."
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "When the conversation began."
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "When the conversation ended, or null if still in progress."
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "LLM-generated one-line gist; null until the conversation ends."
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "description": "Number of messages in the conversation."
          }
        },
        "type": "object",
        "required": [
          "id",
          "started_at",
          "message_count"
        ],
        "title": "ConversationSummaryItem",
        "description": "One row in the history list view.\n\nMirrors :class:`core.conversations.ConversationSummary`. We project\nexplicitly through pydantic instead of returning the dataclass so\nthe OpenAPI schema (and thus the generated TypeScript types) carry\nfield-level descriptions."
      },
      "ConversationsListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationSummaryItem"
            },
            "type": "array",
            "title": "Conversations",
            "description": "Conversations ordered by started_at descending."
          }
        },
        "type": "object",
        "title": "ConversationsListResponse",
        "description": "Paginated response for ``GET /conversations``."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "description": "Liveness probe payload."
      },
      "LivekitTokenRequest": {
        "properties": {
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room",
            "description": "Optional LiveKit room name; defaults to `user-{userId}`."
          }
        },
        "type": "object",
        "title": "LivekitTokenRequest",
        "description": "Optional body for `/livekit/token`.\n\nThe room is optional \u2014 when omitted, the API picks the per-user\ndefault (``user-{userId}``). Letting the client name a room is\nuseful for tests and for future multi-room features but is not\nrequired by the demo."
      },
      "LivekitTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Short-lived LiveKit access JWT."
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "LiveKit server URL the browser should dial."
          },
          "room": {
            "type": "string",
            "title": "Room",
            "description": "Room name encoded in the token's grants."
          }
        },
        "type": "object",
        "required": [
          "token",
          "url",
          "room"
        ],
        "title": "LivekitTokenResponse",
        "description": "Connection bundle returned to the browser."
      },
      "MeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "MeResponse",
        "description": "Authenticated user, projected for the wire."
      },
      "MessageItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "One of 'user', 'assistant', 'tool'."
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          },
          "tool_args": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Args"
          },
          "tool_result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Tool Result"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "MessageItem",
        "description": "A single transcript turn for the detail view."
      },
      "PreferenceUpsertRequest": {
        "properties": {
          "value": {
            "title": "Value",
            "description": "New value for the preference."
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "PreferenceUpsertRequest",
        "description": "Body for ``PUT /preferences/{key}``.\n\nA single ``value`` field rather than free-form JSON so the OpenAPI\nschema (and thus the generated TS types) carries an explicit shape\nthe frontend can typecheck. The field is ``Any``-typed to allow\nthe existing structured-preferences contract (string today,\nstructured values later); validation against the recognised key\ncatalogue happens server-side via\n:func:`core.preferences.validate_preference`."
      },
      "PreferenceUpsertResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "PreferenceUpsertResponse",
        "description": "Echo of the stored value after a ``PUT`` succeeds."
      },
      "PreferencesResponse": {
        "properties": {
          "preferences": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preferences",
            "description": "All stored preferences for the authenticated user."
          }
        },
        "type": "object",
        "title": "PreferencesResponse",
        "description": "Flat key-value map of the authenticated user's preferences.\n\nThe shape is intentionally a single ``preferences`` field rather\nthan the response body itself being a free-form object \u2014 FastAPI\n+ OpenAPI handle nested objects more reliably than top-level\n``additionalProperties`` schemas, and the frontend reads\n``response.preferences`` either way."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}